chore: restructure project layout and move tests to /tests
