+++ Cargo.toml
@@ -30,4 +30,5 @@ serde_json = "1.0"
 slog-scope = "4"
 smart-default = "0.7"
 tokio = { version = "1", features = ["full"] }
+tokio-test = "0.4"  # Explicitly add if missing
 ureq = { version = "3", features = ["json"] }
